実装/LCD Fixes #95
