+++ whack-a-mole.X/header/LCD.h
@@ -27,8 +27,6 @@
 #define LINE_DIGITS_MAX 8
 #define LCD_SET_POS_DB7 0x80
 
-// 10進数の最大の数
-#define DECIMAL_MAX 9
 // ClearDisplay コマンドのデータ部
 #define CMD_LCD_CLR_DISPLAY 0x01
 
@@ -59,8 +57,6 @@ void Write1LineToLCD(uint8_t *i_str, uint8_t i_len);
 void ClrLineDisplay(uint8_t i_line);
 void ClrDisplay(void);
 
-void ItoStr(uint16_t i_value, uint8_t *o_strAdd, uint8_t i_strLen);
-
 // LCDResetFlg
 // LCDのリセット処理を、このリセット処理が終わってから行うようにするためのフラグ
 inline void SetLCDResetFlg(void);  // ON
@@ -72,7 +68,5 @@ void DisplayOFF(void);
 extern bool UpdateLCDFlg;
 // LCDのリセット処理を、このリセット処理が終わってから行うようにするためのフラグ
 extern bool LCDResetFlg;
-
-char *utoa(unsigned int value, char *s, int radix);
 
 #endif /* LCDCLASS_H */
