+++ whack-a-mole.X/nbproject/configurations.xml
@@ -56,6 +56,7 @@
       <itemPath>src/Rand.c</itemPath>
       <itemPath>src/LCD.c</itemPath>
       <itemPath>src/LED.c</itemPath>
+      <itemPath>src/Common.c</itemPath>
     </logicalFolder>
     <logicalFolder name="ExternalFiles"
                    displayName="Important Files"
